Moved Map Source menu to MapFragment
