+++ app/src/main/res/menu/menu_main.xml
@@ -33,10 +33,6 @@
         android:title="@string/preferences_language"
         app:showAsAction="withText" />
     <item
-        android:id="@+id/preferences_map_style"
-        android:title="@string/preferences_map_style"
-        app:showAsAction="withText" />
-    <item
         android:id="@+id/show_intro"
         android:title="@string/show_intro"
         app:showAsAction="withText" />
